Move everything verticle related to the io.vertx.core.impl.verticle package
