+++ src/main/java/io/vertx/core/impl/verticle/JavaVerticleFactory.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.core.impl;
+package io.vertx.core.impl.verticle;
 
 import io.vertx.core.Promise;
 import io.vertx.core.Verticle;
